+++ lib/pop3mail/header.ex
@@ -30,7 +30,7 @@ defmodule Pop3mail.Header do
    * `header_list` - list with tuples {:header, header name, value}
    * `unsafe_addition` - append this to the filename if the filesystem allows it.
    """
-   @spec store(list({:header, String.t, String.t}), String.t, String.t, String.t) :: {:ok, String.t} | {:error, String.t, String.t}
+   @spec store(list({:header, String.t, String.t}), String.t, String.t, String.t) :: {:ok, String.t} | {:error, atom, String.t}
    def store(header_list, filename_prefix, filename_addition, dirname) do
       date    = lookup(header_list, "Date")
       # RFC 2047, search for encoded words and put these in decoded text list, like  [{charset1,content1},{charset2,content2}]